Type of renderer can now be specified in engine initiallization

+++ DDM3-Lite-Engine/DDM3-Lite/DataTypes/DescriptorObjects/TextureDescriptorObject.cpp
@@ -68,7 +68,7 @@ void DDM3::TextureDescriptorObject::Clear()
 	SetupImageInfos();
 }
 
-int DDM3::TextureDescriptorObject::GetTextureAmount() const
+size_t DDM3::TextureDescriptorObject::GetTextureAmount() const
 {
 	return m_Textures.size();
 }
